sagan_workflow/pyproject.toml
[tool.poetry]
name = "spaider-agent-template"
version = "0.1.0"
description = ""
authors = ["Ketan Kunkalikar <ketankunkalikar33@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
langchain = "*"
langchain-huggingface = "*"
langchain-community = "*"
typing-extensions = "*"
langgraph = "*"
python-dotenv = "*"
openai = "*"
langchain-openai = "*"
langgraph-checkpoint-sqlite = "*"
black = "*"
ipython = "*"
langchain-groq = "*"
chromadb = "^0.5.15"
langchain-chroma = "^0.1.4"
markdown = "^3.7"
pdfkit = "^1.0.0"
langchain-core = "^0.3.16"
duckduckgo-search = "^6.3.4"
pypdf = "^5.1.0"
python-multipart = "^0.0.17"
pyngrok = "^7.2.1"
nest-asyncio = "^1.6.0"
beautifulsoup4 = "^4.12.3"
fitz = "^0.0.1.dev2"
pymupdf = "1.22.5"
einops = "^0.8.0"
sentence-transformers = "^3.3.1"
pylatex = "^1.4.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"